backend/requirements.txt
pyserial==3.5
firebase-admin==6.5.0
